+++ src/customer_service_agent/__init__.py
@@ -8,6 +8,7 @@ from .feedback import load_feedback, replay_feedback, write_feedback_replay
 from .privacy_evaluation import evaluate_redaction_cases, write_redaction_report
 from .classification import LocalLanguageClassificationAdapter
 from .review_export import build_review_report_export
+from .review_history import summarize_review_history
 
 __all__ = [
     "CustomerServiceAgent",
@@ -27,5 +28,6 @@ __all__ = [
     "write_redaction_report",
     "LocalLanguageClassificationAdapter",
     "build_review_report_export",
+    "summarize_review_history",
 ]
-__version__ = "0.8.0"
+__version__ = "0.9.0"